Adding notification methods in DAO

+++ db/migrate/20091219172254_create_notifications.rb
@@ -2,7 +2,7 @@ class CreateNotifications < ActiveRecord::Migration
   def self.up
     create_table :notifications do |t|
       t.string :reason
-      t.decimal :user_id
+      t.integer :user_id
       t.timestamp :date
       t.decimal :object_id
